OPENIG-9436 JwtReconstruction octet-sequence signed JWT support - Change to reduce the extension to only what is required for octet-sequence SignedJwt (JWS) reconstruction. - Renamed OctetSequenceJwsReconstruction.
